PARTIE 1 finieee
happy

+++ Arete.h
@@ -15,8 +15,8 @@ class Arete
         Arete(int idArete, int id1, int id2, float poids1, float poids2);
         int getId1();
         int getId2();
-        int getPoids1();
-        int getPoids2();
+        float getPoids1();
+        float getPoids2();
         void afficher(float poids);
         ~Arete();
